requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.32.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-ollama==0.3.10
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.2.1
numpy==2.0.2
ollama==0.6.0
openai==2.3.0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyparsing==3.2.5
PyPDF2==3.0.1
python-dotenv==1.1.1
python-telegram-bot==22.5
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
